feat(request): modify post google sheet location
