+++ components/MLogin.tsx
@@ -6,6 +6,7 @@ import * as WebBrowser from 'expo-web-browser';
 import * as Google from 'expo-auth-session/providers/google';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import axios from 'axios';
+import { postGoogleSheets } from '../logic/requests/post_google_sheets';
 
 WebBrowser.maybeCompleteAuthSession();
 
@@ -49,57 +50,12 @@ const MLogin = ({navigation}) => {
     }
   }
 
-  const appendDataToGoogleSheet = async () => {
-    const accessToken = await AsyncStorage.getItem('accessToken');
-    console.log(`accessToken: ${accessToken}`)
-    if (!accessToken) {
-      console.error("No access token found");
-      return;
-    }
-
-    const sheetId = '12P5-URZO0SfO5-ebVALqgCohdaEhcjWeFkDgAzJmNwY';
-    const range = 'DATOS!A:A';
-    const url = `https://sheets.googleapis.com/v4/spreadsheets/${sheetId}/values/${range}:append?valueInputOption=RAW`;
-
-    const body = {
-      majorDimension: "ROWS",
-      values: [
-        ["2023-12-12", "Valor desde Android!!!"]
-      ]
-    };
-
-    let data = {"majorDimension": "ROWS",  "values": [["2023-03-01","Lo logre de ANDROOIDDD"]]};
-
-    let config = {
-      method: 'post',
-      maxBodyLength: Infinity,
-      url: 'https://sheets.googleapis.com/v4/spreadsheets/12P5-URZO0SfO5-ebVALqgCohdaEhcjWeFkDgAzJmNwY/values/DATOS!A:A:append?valueInputOption=RAW',
-      headers: {
-        'Content-Type': 'application/json', 
-        'Authorization': 'Bearer ' + accessToken
-      },
-      data : JSON.stringify(data)
-    };
-
-    axios.request(config)
-    .then((response) => {
-      console.log(JSON.stringify(response.data));
-    })
-    .catch((error) => {
-      console.log(error);
-      console.log(error.response.status);
-      console.log(error.response.data);
-    });
-
-  };
-
-
   return (
     <View style={styles.container}>
       <Text style={styles.description}>{JSON.stringify(userInfo, null, 2)}</Text>
       <Text style={styles.title}>Metrics Uploader</Text>
       <MButton type="primary" text="Sign in with Google" onPress={() => promptAsync()} />
-      <MButton type="primary" text="Append to google sheets" onPress={appendDataToGoogleSheet} />
+      <MButton type="primary" text="Append to google sheets" onPress={postGoogleSheets} />
       <MButton type="primary" text="Delete local storage" onPress={() => AsyncStorage.removeItem("@user")} />
     </View>
   );
